+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 #### :bug: Bug fix
 
 - Editor: Fix issue where pipe completions would not trigger with generic type arguments. https://github.com/rescript-lang/rescript/pull/7231
+- Fix leftover assert false in code for `null != undefined`. https://github.com/rescript-lang/rescript/pull/7232
 
 # 12.0.0-alpha.7